+++ src/layouts/MainLayout.vue
@@ -32,13 +32,6 @@
           <q-btn
             class="col"
             icon="ballot"
-            label="Missions"
-            no-wrap no-caps
-            to="/missions"
-          />
-          <q-btn
-            class="col"
-            icon="ballot"
             label="Asset Classes"
             no-wrap no-caps
             to="/assetclasses"
@@ -60,10 +53,6 @@
         <q-item to="/executors">
           <q-item-side icon="settings"/>
           <q-item-main label="Executors"/>
-        </q-item>
-        <q-item to="/missions">
-          <q-item-side icon="ballot"/>
-          <q-item-main label="Missions"/>
         </q-item>
         <q-item to="/assetclasses">
           <q-item-side icon="ballot"/>